fix(card): fix product cards layout and responsiveness
This commit fixes the product cards layout in all pages, where the cards in some pages used to have the previous design (4-cards layout), to update all the sections to have updated layout design (5-cards layout) in desktop devices. Additionally, it modifies the cards layout in different devices to fix responsiveness issues, such as not aligining in a single row (for tablet) and too many cards for mobile devices.