fix(useSearchParams): fix prerendering/suspense error
This commit fixes the errors caused by useSearchParams in all the components, by wrapping the components using React's Suspense. Additionally, a loading component for UX and a loading spinner component is created for resuasability. Finally, the same issue for the restricted parts of websites are temporarily solved by wrapping the layout component inside dashboard inside <Suspense>.